Fix failing infrastructure test
